+++ sdk/monitor/azquery/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.1.1 (Unreleased)
+## 1.2.0-beta.2 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,14 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.2.0-beta.1 (2023-11-16)
+
+### Features Added
+* Added `MetricsBatchClient` to support batch querying metrics from Azure resources
+
+### Other Changes
+* Enabled spans for distributed tracing.
 
 ## 1.1.0 (2023-05-09)